refactor(sql-gen): move clickhouse_query_generator into sql_generator/emitters/clickhouse
Phase 0.3 of the DeltaGraph plan. The ClickHouse SQL emitter (17 files,
~16.7K LoC) moves from a top-level crate module to its proper home
under the dialect abstraction tree, alongside the SqlEmitter and
FunctionMapper traits introduced in Phase 0.1/0.2. The planned
Databricks emitter (Phase 1) will land as a sibling at
sql_generator/emitters/databricks/.

The old top-level path is preserved as a transparent re-export in
lib.rs:

    pub use crate::sql_generator::emitters::clickhouse
        as clickhouse_query_generator;

so the ~35 internal call sites that say
`crate::clickhouse_query_generator::X` keep working unchanged. They can
migrate to the new path at their own pace; nothing forces the churn now.

Implementation: a single `git mv` of the directory, plus two-line
edits to lib.rs and sql_generator/mod.rs, plus a one-line
emitters/mod.rs. All 18 files are tracked as pure renames (zero content
diff) so the PR is mechanically reviewable.

Verified: cargo build, cargo fmt --all, cargo clippy --all-targets --
-D warnings all clean; full test suite green (1344 lib + 279
integration + 36 misc, zero failures — same totals as PRs #316 / #317).
