+++ src/sql_generator/emitters/clickhouse/function_registry.rs
