+++ src/sql_generator/emitters/clickhouse/view_scan.rs
